requirements.txt
pymongo==3.11.2
dnspython==2.0.0
discord.py==1.5.1
requests>=2.24.0
flask>=1.1.2
PyNaCl>=1.4.0
python-dotenv>=0.18.0
yagmail>=0.14.256
cn2an>=0.5.11